requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.51.3
pandas==2.2.2
accelerate==0.31.0
sentence-transformers==4.1.0
datasets==2.20.0
aiohttp==3.9.5
evaluate==0.4.3
nltk==3.9.1
faiss-cpu==1.7.4
mauve-text==0.4.0
numpy==1.26.4